implement new integration page layout using the SideNav component and add MDM
