+++ frontend/context/app.tsx
@@ -153,8 +153,7 @@ const setPermissions = (
     isSandboxMode: permissions.isSandboxMode(config),
     isFreeTier: permissions.isFreeTier(config),
     isPremiumTier: permissions.isPremiumTier(config),
-    // isMdmEnabled: permissions.isMdmEnabled(config),
-    isMdmEnabled: true, // MDM TODO: Remove when backend is merged
+    isMdmEnabled: permissions.isMdmEnabled(config),
     isGlobalAdmin: permissions.isGlobalAdmin(user),
     isGlobalMaintainer: permissions.isGlobalMaintainer(user),
     isGlobalObserver: permissions.isGlobalObserver(user),
