+++ frontend/pages/admin/AppSettingsPage/OrgSettingsNavItems.tsx
@@ -1,5 +1,3 @@
-import React from "react";
-
 import PATHS from "router/paths";
 
 import { ISideNavItem } from "../components/SideNav/SideNav";
@@ -20,56 +18,56 @@ const ORG_SETTINGS_NAV_ITEMS: ISideNavItem<IAppConfigFormProps>[] = [
     title: "Organization info",
     urlSection: "organization",
     path: PATHS.ADMIN_SETTINGS_INFO,
-    Card: (props) => <Info {...props} />,
+    Card: Info,
   },
   {
     title: "Fleet web address",
     urlSection: "webaddress",
     path: PATHS.ADMIN_SETTINGS_WEBADDRESS,
-    Card: (props) => <WebAddress {...props} />,
+    Card: WebAddress,
   },
   {
     title: "Single sign-on options",
     urlSection: "sso",
     path: PATHS.ADMIN_SETTINGS_SSO,
-    Card: (props) => <Sso {...props} />,
+    Card: Sso,
   },
   {
     title: "SMTP options",
     urlSection: "smtp",
     path: PATHS.ADMIN_SETTINGS_SMTP,
-    Card: (props) => <Smtp {...props} />,
+    Card: Smtp,
   },
   {
     title: "Agent options",
     urlSection: "agents",
     path: PATHS.ADMIN_SETTINGS_AGENTS,
-    Card: (props) => <Agents {...props} />,
+    Card: Agents,
   },
   {
     title: "Host status webhook",
     urlSection: "host-status-webhook",
     path: PATHS.ADMIN_SETTINGS_HOST_STATUS_WEBHOOK,
-    Card: (props) => <HostStatusWebhook {...props} />,
+    Card: HostStatusWebhook,
   },
   {
     title: "Usage statistics",
     urlSection: "statistics",
     path: PATHS.ADMIN_SETTINGS_STATISTICS,
-    Card: (props) => <Statistics {...props} />,
+    Card: Statistics,
   },
   {
     title: "Fleet Desktop",
     urlSection: "fleet-desktop",
     path: PATHS.ADMIN_SETTINGS_FLEET_DESKTOP,
     // isPremium: true,
-    Card: (props) => <FleetDesktop {...props} />,
+    Card: FleetDesktop,
   },
   {
     title: "Advanced options",
     urlSection: "advanced",
     path: PATHS.ADMIN_SETTINGS_ADVANCED,
-    Card: (props) => <Advanced {...props} />,
+    Card: Advanced,
   },
 ];
