+++ frontend/pages/admin/AppSettingsPage/_styles.scss
@@ -1,5 +1,4 @@
 .app-settings {
-  padding: $pad-xxlarge 30px;
 
   &__page-description {
     font-size: $x-small;
