+++ frontend/pages/admin/IntegrationsPage/IntegrationPage.tests.tsx
@@ -1,0 +1,21 @@
+import React from "react";
+import { screen } from "@testing-library/react";
+
+import { createCustomRenderer } from "test/test-utils";
+import IntegrationsPage from "./IntegrationsPage";
+
+describe("Integrations Page", () => {
+  it("renders the MDM section in the side nav if MDM feature is enabled", () => {
+    const render = createCustomRenderer({
+      context: {
+        app: { isMdmEnabled: true },
+      },
+    });
+
+    render(<IntegrationsPage params={{ section: "mdm" }} />);
+
+    expect(
+      screen.getByText("Mobile Device Management (MDM)")
+    ).toBeInTheDocument();
+  });
+});
