+++ frontend/pages/admin/IntegrationsPage/cards/Integrations/IntegrationsTableConfig.tsx
@@ -10,8 +10,8 @@ import {
 } from "interfaces/integration";
 import { IDropdownOption } from "interfaces/dropdownOption";
 
-import JiraIcon from "../../../../assets/images/icon-jira-24x24@2x.png";
-import ZendeskIcon from "../../../../assets/images/icon-zendesk-32x24@2x.png";
+import JiraIcon from "../../../../../../assets/images/icon-jira-24x24@2x.png";
+import ZendeskIcon from "../../../../../../assets/images/icon-zendesk-32x24@2x.png";
 
 interface IHeaderProps {
   column: {
