+++ frontend/pages/admin/IntegrationsPage/cards/Integrations/components/DeleteIntegrationModal/DeleteIntegrationModal.tsx
