+++ frontend/pages/admin/IntegrationsPage/cards/Integrations/components/DeleteIntegrationModal/_styles.scss
