+++ frontend/pages/admin/IntegrationsPage/cards/Integrations/components/DeleteIntegrationModal/index.ts
