+++ frontend/pages/admin/IntegrationsPage/cards/Integrations/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./Integrations";
