+++ frontend/pages/admin/SettingsWrapper/_styles.scss
@@ -33,10 +33,6 @@
     }
   }
 
-  .app-settings {
-    padding: 0;
-  }
-
   .component__tabs-wrapper {
     top: $pad-xxlarge; // for sticky
     z-index: 3;
